requirements.txt
discord.py==2.3.2
emoji==2.10.0
pandas==2.2.3
Pillow==11.0.0
python-dotenv==1.0.1
Requests==2.32.3
